test(providers): narrow readiness support ownership
