+++ .github/ci-impact.json
@@ -344,7 +344,6 @@
         "tests/support/_agent_ab_support.py",
         "tests/support/commands.py",
         "tests/support/domains.py",
-        "tests/support/providers.py",
         "tests/support/services.py",
         "tests/support/state.py",
         "tests/conftest.py"
