+++ tests/boundary/providers/flint/runtime/test_linear_rational_solution.py
@@ -10,8 +10,8 @@ from typing import Any
 
 import pytest
 from tests.component.capabilities.capabilities import invoke_capability as _invoke
-from tests.support.services import open_domain_services
 from tests.support.rationals import rational_payload as _q
+from tests.support.services import open_domain_services
 
 import jacobian.provider_runtime as provider_runtime
 from jacobian.bounded_process import BoundedProcessResult
