+++ tests/boundary/providers/lean/test_lean_proof_edit.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import pytest
-from tests.support.providers import (
+from tests.support.provider_lean import (
     PINNED_MATHLIB_RUNTIME_UNAVAILABLE_REASON,
     pinned_mathlib_runtime_available,
 )
