+++ tests/boundary/providers/lean/test_lean_statement_capabilities.py
@@ -6,7 +6,7 @@ import shutil
 from pathlib import Path
 
 import pytest
-from tests.support.providers import (
+from tests.support.provider_lean import (
     PINNED_MATHLIB_RUNTIME_UNAVAILABLE_REASON,
     pinned_mathlib_runtime_available,
 )
