+++ tests/component/checkers/test_exact_probability_operation_checkers.py
@@ -5,9 +5,9 @@ from collections.abc import Callable
 from typing import Any
 
 import pytest
+from tests.support.rationals import rational_payload as _q
 from tests.unit.contracts.artifacts import artifact_uri as _uri
 from tests.unit.contracts.artifacts import canonical_digest as _digest
-from tests.support.rationals import rational_payload as _q
 
 from jacobian_checkers.exact_probability_operations import (
     check_finite_condition,
