+++ tests/component/providers/matrix/test_linear_rational_inconsistency.py
@@ -8,8 +8,8 @@ from typing import Any
 
 import pytest
 from tests.component.capabilities.capabilities import invoke_capability as _invoke
-from tests.support.services import open_domain_services
 from tests.support.rationals import rational_payload as _q
+from tests.support.services import open_domain_services
 
 from jacobian.bounded_process import BoundedProcessResult
 from jacobian.contracts.capabilities import (
