+++ tests/e2e/capability_workflows/test_capability_round_trips.py
@@ -7,7 +7,7 @@ from typing import Any
 
 import pytest
 from mcp import Client
-from tests.support.providers import (
+from tests.support.provider_lean import (
     PINNED_MATHLIB_RUNTIME_UNAVAILABLE_REASON,
     pinned_mathlib_runtime_available,
 )
